Game screen!
added two scripts in <head> as it is not a rendering component .

First script is the CDN for Phaser (our jscript framework), second script is the CDN for an external plugin that will determine our physics engine, which is a RigidBody one, think collision based events. eg. when a ball impacts a table, a sound is produced the sound is the event and the ball and table are objects.